kadai15

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,26 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
+        
+        return view('welcome', $data);
     }
+
+        
+    
 
     /**
      * Show the form for creating a new resource.
@@ -49,11 +63,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task= new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            
+        ]);
         return redirect('/');
     }
 
@@ -118,7 +132,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+    
 
         return redirect('/');
     }
